requirements.txt
torchvision==0.4.1
numpy==1.17.2
opencv_python==4.1.1.26
torch==1.3.0
Pillow==6.2.1
